+++ multiversx_sdk_core/transaction_factories/token_operations_factory_test.py
@@ -1,9 +1,9 @@
 
 
 from multiversx_sdk_core import Address
-from multiversx_sdk_core.token_operations.token_operations_config import \
+from multiversx_sdk_core.transaction_factories.token_operations_config import \
     TokenOperationsConfig
-from multiversx_sdk_core.token_operations.token_operations_factory import \
+from multiversx_sdk_core.transaction_factories.token_operations_factory import \
     TokenOperationsFactory
 
 frank = Address.from_bech32("erd1kdl46yctawygtwg2k462307dmz2v55c605737dp3zkxh04sct7asqylhyv")
